Factor out logic from main function
